+++ Chapter.Net.WPF.Controls/ChapterNumberBox/ChapterNumberBoxSelection.cs
@@ -1,5 +1,5 @@
 // -----------------------------------------------------------------------------------------------------------------
-// <copyright file="NumberBoxSelection.cs" company="my-libraries">
+// <copyright file="ChapterNumberBoxSelection.cs" company="my-libraries">
 //     Copyright (c) David Wendland. All rights reserved.
 // </copyright>
 // -----------------------------------------------------------------------------------------------------------------
@@ -9,9 +9,9 @@
 namespace Chapter.Net.WPF.Controls
 {
     /// <summary>
-    ///     Defines how the automatic selection of the number in the <see cref="NumberBox" /> should behave.
+    ///     Defines how the automatic selection of the number in the <see cref="ChapterNumberBox" /> should behave.
     /// </summary>
-    public enum NumberBoxSelection
+    public enum ChapterNumberBoxSelection
     {
         /// <summary>
         ///     No automatic selection will be done.
@@ -19,7 +19,7 @@ namespace Chapter.Net.WPF.Controls
         None,
 
         /// <summary>
-        ///     The number gets selected when the NumberBox got the focus.
+        ///     The number gets selected when the ChapterNumberBox got the focus.
         /// </summary>
         OnFocus,
 
@@ -29,7 +29,7 @@ namespace Chapter.Net.WPF.Controls
         OnUpDown,
 
         /// <summary>
-        ///     The number will be selected when the NumberBox got the focus or the value gets incremented or decremented using the
+        ///     The number will be selected when the ChapterNumberBox got the focus or the value gets incremented or decremented using the
         ///     arrow keys or up/down buttons.
         /// </summary>
         OnFocusAndUpDown