+++ Chapter.Net.WPF.Controls/ChapterNumberBox/NumberChangedEventHandler.cs
@@ -1,5 +1,5 @@
 // -----------------------------------------------------------------------------------------------------------------
-// <copyright file="ValueChangedEventHandler.cs" company="my-libraries">
+// <copyright file="NumberChangedEventHandler.cs" company="my-libraries">
 //     Copyright (c) David Wendland. All rights reserved.
 // </copyright>
 // -----------------------------------------------------------------------------------------------------------------
@@ -9,9 +9,9 @@
 namespace Chapter.Net.WPF.Controls
 {
     /// <summary>
-    ///     The event handler for the <see cref="NumberBox.NumberChanged" /> event.
+    ///     The event handler for the <see cref="ChapterNumberBox.NumberChanged" /> event.
     /// </summary>
-    /// <param name="sender">The corresponding NumberBox.</param>
+    /// <param name="sender">The corresponding ChapterNumberBox.</param>
     /// <param name="e">The object with the old and new number.</param>
     public delegate void NumberChangedEventHandler(object sender, NumberChangedEventArgs e);
 }